filter option and listing stuck renewing subscriptions (#21)
* initial work on having a page to display subscriptions stuck in the renewal process

* added filter by status on the subscriptions listing page

start showing the renewing subscription. these are subscription(state) that were stuck therefore they belong to this list

* decided to just add a filter and add the renewing state in the failure page

* as required by travis when creating the dummy app:
"deface is required to run solidus_auth_devise with solidus versions < 2.5.
Please add deface to your Gemfile"

also time to bump up the gem version. this is now using a state machine
to set the current state of the subscription